### 2019-07-16 #1573 @whedon
[REVIEW]: TimeSeriesClustering: An extensible framework in Julia
**Submitting author:** @holgerteichgraeber (<a href="http://orcid.org/0000-0002-4061-2226">Holger Teichgraeber</a>)
**Repository:** <a href="https://github.com/holgerteichgraeber/TimeSeriesClustering.jl" target ="_blank">https://github.com/holgerteichgraeber/TimeSeriesClustering.jl</a>
**Version:** v0.5.2
