### 2019-07-16 #1573 @danielskatz
Note: @ahwillia is traveling for the next few weeks. It will be about 3 weeks until he can work on this review.
