### 2019-07-23 #1573 @holgerteichgraeber
> 👋 @jgoldfar, @ahwillia - We'll do the review here - please read the comments above, and get started when you can. If you have any questions, please ask.

👋Thank you all for offering to review, I look forward to your comments.
Tagging along co-authors @YoungFaithful and @arbrandt for reference.
