### 2019-07-26 #1573 @danielskatz
👋 @jgoldfar - have you had a chance to get started?
