### 2019-08-01 #1573 @jgoldfar
Yes; I will post my review within a few days

> On Jul 26, 2019, at 12:55 PM, Daniel S. Katz <notifications@github.com> wrote:
>
> 👋 @jgoldfar - have you had a chance to get started?
>
