### 2019-08-09 #1573 @danielskatz
> Yes; I will post my review within a few days

@jgoldfar - any update on this?
