### 2019-08-09 #1573 @ahwillia
Looking through this now. I'm generally very impressed and think we should be able to approve this in short order. I am confirming that I can install and execute the package now.

My biggest piece of feedback is that the README and description of the package should emphasize even more applications. I think this package will be **broadly** useful to many fields! For example, the first sentence of the README might lead users to think the package is for a very specialized purpose I would recommend editing to something like...

**Current:** "ClustForOpt is a julia implementation of unsupervised machine learning methods for finding representative periods for energy systems optimization problems."

**Revised:** "ClustForOpt is a julia implementation of unsupervised machine learning methods for *detecting motifs, clustering, and quantifying similarity between time series datasets*."

Likewise, in the subsequent paragraphs, I recommend adding some more example applications with citations. Segmentation and clustering of audio datasets should be an easy one to find.

It is of course okay to say something like *"this package was **originally** developed for energy systems optimization"* but I think emphasizing the generality of the package and the methods as much as possible will increase the impact of this work.

Full disclosure, I've worked on using very simple time warping methods for neural data (https://www.biorxiv.org/content/10.1101/661165v1), though those data show very different statistics and call for different modeling approaches. But I'm quite enthusiastic about this area of research.
