### 2019-08-09 #1573 @ahwillia
One final thought, I don't insist on changing the name of the repo, but something like "TimeSeriesClustering.jl" would seem to better capture the function of the package. The name ClustForOpt doesn't make it super clear what the package does...

Also can the authors comment on the differences between this package and other time series packages in julia (e.g `TimeSeries.jl`) in the paper / README? It would be nice to give users more guidance on the broader tools available in Julia for these kinds of modeling problems.
