### 2019-08-10 #1573 @holgerteichgraeber
Thank you for your feedback, these are great ideas! I am out for the weekend, and will get back to this next week.
