### 2019-08-10 #1573 @holgerteichgraeber
In case that there are any papers that you can recommend to read in the suggested application areas, suggestions are greatly appreciated.
