### 2019-08-15 #1573 @danielskatz
👋 @jgoldfar, @ahwillia - can you please use your checklists above to indicate what you think is ok, and what needs to be done, in addition to the comments @ahwillia has posted in this thread, and what I expect @jgoldfar to post soon.
