### 2019-08-16 #1573 @ahwillia
Checked my boxes...

> @holgerteichgraeber
>In case that there are any papers that you can recommend to read in the suggested application areas, suggestions are greatly appreciated.

Eamonn Keogh has a variety of methods and application papers to check out (e.g. https://www.cs.ucr.edu/~eamonn/MatrixProfile.html). I'm sure many other research groups have relevant papers as well. Please don't worry about being comprehensive, but the more references you can find the better.
