### 2019-08-17 #1573 @danielskatz
👋 @jgoldfar

> Yes; I will post my review within a few days

Can you please go ahead and do this? (and check the boxes for items that are complete)
