### 2019-08-18 0871d6c
Paper review comments TimeSeriesClustering - generalization
