+++ paper/paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'ClustForOpt: Time-series aggregation for optimization in Julia'
+title: 'TimeSeriesClustering: An extensible framework in Julia'
 tags:
   - Julia
   - clustering
@@ -19,42 +19,83 @@ authors:
 affiliations:
  - name: Department of Energy Resources Engineering, Stanford University
    index: 1
-date: 5 July 2019
+date: 18 August 2019
 bibliography: paper.bib
 ---
 
 # Summary
 
-``ClustForOpt`` is an extensible framework for time-series aggregation in Julia. It is designed specifically to perform time series aggregation for energy systems optimization problems, though could be adapated to be used with arbitrary time series datasets. ``ClustForOpt`` provides a type system for temporal data that can easily be integrated with the formulation of optimization problems, and provides an implementation of the most commonly used clustering methods and extreme value selection methods for temporal data. ``ClustForOpt`` provides simple integration of multiple attributes (e.g. wind availability, solar availability, and electricity demand) in a single aggregation process.
+``TimeSeriesClustering`` is a Julia implementation of unsupervised learning methods for time series data sets. It provides functionality for clustering and aggregating, detecting motifs, and quantifying similarity between time series datasets.
+The software provides a type system for temporal data, and provides an implementation of the most commonly used clustering methods and extreme value selection methods for temporal data.
+``TimeSeriesClustering`` provides simple integration of multi-dimensional time-series data (e.g. multiple attributes such as wind availability, solar availability, and electricity demand) in a single aggregation process.
+The software is applicable to general time series datasets and lends itself well to a multitude of application areas such as INSERT
+``TimeSeriesClustering`` was originally developed to perform time series aggregation for energy systems optimization problems. Because of the software's origin, many of the examples in this work stem from the field of energy systems optimization.
 
-The unique design of ``ClustForOpt`` allows for scientific comparison of the performance of different time-series aggregation methods, both in terms of the statistical error measure and in terms of optimization outcome.
-The choice of temporal modeling, especially of time-series aggregation methods, can have significant impact on overall optimization outcome, which in the end is used to make policy and business decisions. It is thus important to not view time-series aggregation and optimization model formulation as two seperate, consecutive steps, but to integrate time-series aggregation into the overall process of building an optimization model in an iterative manner. ``ClustForOpt`` allows for this iterative integration in a simple way.
+## General package features
 
-``ClustForOpt`` provides two sample optimization problems to illustrate the integration of time-series aggregation and optimization problem formulation through our type system.
-However, it is generally thought to be independent of the application at hand, and others are encouraged to use the package as a base for their own optimization problem formulation.
-The Julia package [``CapacityExpansion``](https://github.com/YoungFaithful/CapacityExpansion.jl) provides a detailed generation and transmission capacity expansion model built upon ``ClustForOpt``, and illustrates its capabilities in conjunction with a complex optimization problem formulation.
+The unique design of ``TimeSeriesClustering`` allows for scientific comparison of the performance of different time-series aggregation methods, both in terms of the statistical error measure and in terms of its impact on the application outcome.
+The clustering methods that are implemented in ``TimeSeriesClustering`` follow the framework presented by @Teichgraeber:2019, and the extreme value selection methods follow the framework presented by @Lindenmeyer:2019. Using these frameworks allows ``TimeSeriesClustering`` to be generally extensible to new aggregation methods in the future.
 
[…]
