### 2019-08-18 487b27b
Update paper intro
