### 2019-08-18 a2239eb
Rename to TimeSeriesClustering (ClustForOpt -> TimeSeriesClustering)
