### 2019-08-18 c8160d4
Update to more general time series data mining.
