### 2019-08-18 #1573 @holgerteichgraeber
>
>
> @whedon generate pdf from branch paper

Ok, that seems not to have worked. Will push to master branch and revisit.

I will be updating the package to reflect its more general applicability and rename to TimeSeriesClustering.jl. I will update here once complete.
