+++ paper/paper.bib
@@ -340,7 +340,7 @@ year = {1997}
 author = {Kamper, Herman and Livescu, Karen and Goldwater, Sharon},
 journal = {IEEE Workshop on Automatic Speech Recognition and Understanding},
 pages = {719--726},
-title = {An embedded segmental k-means model for unsupervised segmentation and clustering of speech}
+title = {An embedded segmental k-means model for unsupervised segmentation and clustering of speech},
 year = {2017}
 }
 
