### 2019-08-18 cdfd8f1
Update logo
