### 2019-08-19 #1573 @holgerteichgraeber
I am in the process of renaming the package. There is no documentation on how to do that with the new package registry in julia, so I opened an issue in the general registry. Will update here as soon as `TimeSeriesClustering.jl` is officially registered.

@danielskatz would the name of this issue and the repository name at the top need to be updated?
