### 2019-08-26 #1573 @danielskatz
Sorry for the delay - I'm on vacation and just scanning things, and missed this.

But the answer is that we can do what's needed - I'm fairly sure that we will need to change the repo URL in this issue, and we will probably change the name of the issue as well.
