### 2019-08-28 8e239d5
Bump version to 0.5.1
