+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"#![ClustForOpt-logo](assets/clust_for_opt_text.svg)-1","page":"Introduction","title":"(Image: ClustForOpt logo)","text":"","category":"section"},{"location":"#","page":"Introduction","title":"Introduction","text":"(Image: ) (Image: ) (Image: Build Status) (Image: codecov)","category":"page"},{"location":"#","page":"Introduction","title":"Introduction","text":"ClustForOpt is a julia implementation of unsupervised machine learning methods for finding representative periods for energy systems optimization problems. By reducing the number of time steps used in the optimization model, using representative periods leads to significant reductions in computational complexity.","category":"page"},{"location":"#","page":"Introduction","title":"Introduction","text":"The package has three main purposes:","category":"page"},{"location":"#","page":"Introduction","title":"Introduction","text":"Provide a simple process of finding representative periods for time-series input data, with implementations of the most commonly used clustering methods and extreme value selection methods.\nProvide an interface between representative period data and optimization problem by having representative period data stored in a generalized type system.\nProvide a generalized import feature for time series, where variable names, attributes, and node names are automatically stored and can then be used in the definition of sets of the optimization problem later.","category":"page"},{"location":"#","page":"Introduction","title":"Introduction","text":"An example energy systems optimization problem that uses ClustForOpt for its input data is the package CapacityExpansion, which implements a scalable generation and transmission capacity expansion problem.","category":"page"},{"location":"#","page":"Introduction","title":"Introduction","text":"The ClustForOpt package follows the clustering framework presented in Teichgraeber and Brandt, 2019. The package is actively developed, and new features are continuously added. For a reproducible version of the methods and data of the original paper by Teichgraeber and Brandt, 2019, please refer to v0.1.","category":"page"},{"location":"#","page":"Introduction","title":"Introduction","text":"This package is developed by Holger Teichgraeber @holgerteichgraeber and Elias Kuepper @YoungFaithful.","category":"page"},{"location":"#Installation-1","page":"Introduction","title":"Installation","text":"","category":"section"},{"location":"#","page":"Introduction","title":"Introduction","text":"This package runs under julia v1.0 and higher. Install using:","category":"page"},{"location":"#","page":"Introduction","title":"Introduction","text":"import Pkg\nPkg.add(\"ClustForOpt\")","category":"page"},{"location":"#Citing-ClustForOpt-1","page":"Introduction","title":"Citing ClustForOpt","text":"","category":"section"},{"location":"#","page":"Introduction","title":"Introduction","text":"If you find ClustForOpt useful in your work, we kindly request that you cite the following paper (link):","category":"page"},{"location":"#","page":"Introduction","title":"Introduction","text":"  @article{Teichgraeber2019,\n  author = {Holger Teichgraeber and Adam Brandt},\n  title = {Clustering methods to find representative periods for the optimization of energy systems: An initial framework and comparison},\n  journal = {Applied Energy},\n  volume = {239},\n  pages = {1283–1293},\n  year = {2019},\n  doi = {https://doi.org/10.1016/j.apenergy.2019.02.012},\n  }","category":"page"},{"location":"quickstart/#Quick-Start-Guide-1","page":"Quick Start Guide","title":"Quick Start Guide","text":"","category":"section"},{"location":"quickstart/#","page":"Quick Start Guide","title":"Quick Start Guide","text":"This quick start guide introduces the main concepts of using ClustForOpt. For more detail on the different functionalities that ClustForOpt provides, p
[…]
