+++ v0.5.1/search_index.js
@@ -1,0 +1,3 @@
+var documenterSearchIndex = {"docs":
+[{"location":"load_data/#Load-Data-1","page":"Load Data","title":"Load Data","text":"","category":"section"},{"location":"load_data/#","page":"Load Data","title":"Load Data","text":"Here, we describe how to load time-series data into the ClustData format for use in TimeSeriesClustering, and we describe how data is stored in ClustData. Data can be loaded from example data sets provided by us, or you can load your own data.","category":"page"},{"location":"load_data/#Load-example-data-from-TimeSeriesClustering-1","page":"Load Data","title":"Load example data from TimeSeriesClustering","text":"","category":"section"},{"location":"load_data/#","page":"Load Data","title":"Load Data","text":"The example data can be loaded using the following function.","category":"page"},{"location":"load_data/#","page":"Load Data","title":"Load Data","text":"load_timeseries_data(::Symbol)","category":"page"},{"location":"load_data/#TimeSeriesClustering.load_timeseries_data-Tuple{Symbol}","page":"Load Data","title":"TimeSeriesClustering.load_timeseries_data","text":"function load_timeseries_data(existing_data::Symbol;\n                          region::String=\"none\",\n                          T::Int=24,\n                          years::Array{Int,1}=[2016],\n                          att::Array{String,1}=Array{String,1}())\n\nReturn time series of example data sets as ClustData struct.\n\nThe choice of example data set is given by e.g. existing_data=:CEP-GER1. Example data sets are:\n\n:DAM_CA : Hourly Day Ahead Market Electricity prices for California-Stanford 2015\n:DAM_GER : Hourly Day Ahead Market Electricity prices for Germany 2015\n:CEP_GER1 : Hourly Wind, Solar, Demand data Germany one node\n:CEP_GER18: Hourly Wind, Solar, Demand data Germany 18 nodes\n\nOptional inputs to load_timeseries_data:\n\nregion-region descriptor\nT- Number of Segments\nyears::Array{Int,1}= The years to be selected from the csv file as specified in years column\natt::Array{String,1}= The attributes to be loaded. If left empty, all attributes will be loaded.\n\n\n\n\n\n","category":"method"},{"location":"load_data/#","page":"Load Data","title":"Load Data","text":"In the following example, we use the function to load the hourly wind, solar, demand data for Germany for 1 node, and the other data can be loaded similarly. Note that more years are available for the two CEP data sets. The data can be found in the data folder.","category":"page"},{"location":"load_data/#","page":"Load Data","title":"Load Data","text":"using TimeSeriesClustering","category":"page"},{"location":"load_data/#","page":"Load Data","title":"Load Data","text":"ts_input_data = load_timeseries_data(:CEP_GER1)","category":"page"},{"location":"load_data/#Load-your-own-data-1","page":"Load Data","title":"Load your own data","text":"","category":"section"},{"location":"load_data/#","page":"Load Data","title":"Load Data","text":"You can also load your own data. Use the load_timeseries_data function and specify the path where the data is located at (either a folder or the filename).","category":"page"},{"location":"load_data/#","page":"Load Data","title":"Load Data","text":"load_timeseries_data(::String)","category":"page"},{"location":"load_data/#TimeSeriesClustering.load_timeseries_data-Tuple{String}","page":"Load Data","title":"TimeSeriesClustering.load_timeseries_data","text":"function load_timeseries_data(data_path::String;\n                          region::String=\"none\",\n                          T::Int=24,\n                          years::Array{Int,1}=[2016],\n                          att::Array{String,1}=Array{String,1}())\n\nReturn all time series as ClustData struct that are stored as csv files in the specified path.\n\nLoads *.csv files in the folder or the file data_path\nLoads all a
[…]
