### 2019-09-01 #1573 @holgerteichgraeber
>
>
> > @arfon - note that this has been renamed - do we need to do anything to tell whedon this, or in an internal database?
>
> The repository has been renamed or the paper, or both?

Both the paper and the repository (and thus also the link to the repository: https://github.com/holgerteichgraeber/TimeSeriesClustering.jl )
