### 2019-09-01 91c58d5
minor grammar changes
