### 2019-09-01 #1573 @danielskatz
I've suggested some minor grammar changes in https://github.com/holgerteichgraeber/TimeSeriesClustering.jl/pull/118
