### 2019-09-01 #1573 @danielskatz
Also, please make an archive of the repo in an archival repository (e.g., zenodo), and let me know the DOI of the archive, as well as the current version number
