### 2019-09-02 85a41a8
Update README and docs to reflect general applicability of TimeSeriesClustering. Update references.
