### 2019-09-04 #1573 @holgerteichgraeber
The current version number is `v0.5.2` and the zenodo DOI is `10.5281/zenodo.3385349` .
