### 2019-09-04 64c9a0f
Update paper reference
