+++ paper/paper.bib
@@ -14,9 +14,9 @@ year = {2019}
 
 @article{Lindenmeyer:2020,
 author = {Lindenmeyer, Constantin and Teichgraeber, Holger and Baumgaertner, Nils and Kotzur, Leander and Robinius, Martin and Bardow, Andre and Brandt, Adam R},
-journal = {Energy},
+journal = {Energy (Manuscript in preparation)},
 title = {{Extreme events as part of representative periods for the optimization of residential energy supply systems}},
-year = {2020, Manuscript in preparation}
+year = {2020}
 }
 
 @article{Novikov:2019,
