### 2019-09-06 #1573 @holgerteichgraeber
>
>
> The current version number is `v0.5.2` and the zenodo DOI is `10.5281/zenodo.3385349` (https://doi.org/10.5281/zenodo.3385349).

@danielskatz just following up, this may have gotten lost among the other messages.
