### 2019-09-06 #1573 @danielskatz
👋 @holgerteichgraeber - It looks like the two missing DOIs found above are correct - please add them.  And the invalid DOI is indeed invalid - either replace it with the right one or remove it.

Then please regenerate the pdf.
