### 2019-09-07 af20b39
Update dois in paper
