+++ paper/paper.bib
@@ -311,6 +311,7 @@ year = {2014}
 
 @article{Castro:2010,
 author = {Castro, Nuno and Azevedo, Paulo},
+doi = {https://doi.org/10.1137/1.9781611972801.73},
 journal = {Proceedings of the 10th SIAM International Conference on Data Mining, SDM 2010},
 keywords = {Frequent patterns,Motif discovery,Multiresolution,Time series},
 pages = {665--676},
@@ -338,6 +339,7 @@ year = {1997}
 }
 @article{Kamper:2017,
 author = {Kamper, Herman and Livescu, Karen and Goldwater, Sharon},
+doi = {https://doi.org/10.1109/asru.2017.8269008},
 journal = {IEEE Workshop on Automatic Speech Recognition and Understanding},
 pages = {719--726},
 title = {An embedded segmental k-means model for unsupervised segmentation and clustering of speech},
@@ -368,7 +370,6 @@ year = {2007}
 
 @article{Lin:2002,
 author = {Lin, Jessica and Keogh, Eamonn and Lonardi, Stefano and Patel, Pranav},
-doi = {10.1.1.19.6629},
 journal = {Proc. of the 2nd Workshop on Temporal Data Mining},
 pages = {53--68},
 title = {{Finding motifs in time series}},
