### 2019-09-07 #1573 @holgerteichgraeber
>
>
> 👋 @holgerteichgraeber - It looks like the two missing DOIs found above are correct - please add them. And the invalid DOI is indeed invalid - either replace it with the right one or remove it.
>
> Then please regenerate the pdf.

@danielskatz Thank you, I updated the two DOIs and removed the third.
